requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
chromadb
pandas
tqdm
tiktoken
python-dotenv